// src/types/products.tsx
import { StaticImageData } from "next/image";

export type Product = {
  title: string;
  description: string;
  thumbnail: StaticImageData | string;
  images: (StaticImageData | string)[];
  href: string;
  slug?: string;
  stack?: string[];
  content?: React.ReactNode | string;
  video?: string;
};